feat: 필수 조회기능

+++ src/main/java/com/example/shoppingmall/domain/store/controller/StoreController.java
@@ -22,7 +22,7 @@ public class StoreController {
 
 	@GetMapping
 	public ResponseEntity<List<Store>> getStores(
-		@RequestParam(required = false) Double rating,
+		@RequestParam(required = false) Long rating,
 		@RequestParam(required = false) String status) {
 
 		List<Store> stores = storeService.getFilteredStores(rating, status);
